Remember results across nixpkgs versions
The downside of this is that it might miss some problems where a build
is sometimes reproducible but not always. To catch those (e.g. on CI)
you'd want to remove any previous `reproducibility-log.json` before
running another build.

The advantage of this is that it becomes much faster to generate
reports for different nixpkgs versions that are fairly close
together (since many derivations are actually identical and don't
have to be re-checked).

A smaller advantage is that (by manually renaming and fixing
reproducibility-log.json.part) you can now resume an aborted
report.